% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecc_count.R
\name{pecc_count}
\alias{pecc_count}
\title{Statistique of a boolean covariate, analyze by categorical covariates.}
\usage{
pecc_count(df, metric, col_cov = NULL, row_cov = NULL, str_input = F)
}
\arguments{
\item{df}{Name of the dataframe to analyse}

\item{metric}{Name of the column metric (NSA)}

\item{col_cov}{Name of the first column}

\item{row_cov}{Name of the second column}

\item{str_input}{If true, metric, col_cov and row_cov argument must be strings. If False (default), provide expressions (unquoted arguments)}
}
\value{

}
\description{
We often have a boolean column (responder = 1, non responder = 0 for instance). Pecc_count create a table with the number of
occurence (value = 1) divided by the number of total patients with or without specific categorical covariate (two can be simultaneously analyzed)
}
\examples{
pecc_count(mtcars,metric = vs, col_cov = cyl, row_cov = carb)
}
\author{
Thibaud Derippe (\email{Thibaud.Derippe@gmail.com})
}
