% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecc_table1.R
\name{pecc_table1_original}
\alias{pecc_table1_original}
\title{Table1 packagecreation}
\usage{
pecc_table1_original(
  df,
  ...,
  col1 = NA,
  col2 = NA,
  reduceBy = NA,
  outputExpr = F
)
}
\arguments{
\item{df}{dataset to analyze}

\item{...}{name of the columns to analyze, optionally with "_cont" or "_cat" to force a continuous or categorical behavior. If ... is empty, analyze every columns (using reduce by to extract ID).}

\item{col1}{optional name of a categorical covariate to use as first column}

\item{col2}{optional name of a categorical covariate to use as second column}

\item{reduceBy}{optional name of the columns to reduce the dataset by, separeated by "+"  (eg ID + occ)}

\item{outputExpr}{If F return directly the table1 output, if "Both" return both the code and its evaluation, else return only the code}
}
\description{
Re compute table1 from sratch
}
\author{
Thibaud Derippe (\email{Thibaud.Derippe@gmail.com})
}
