% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_boxplot.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Boxplot simplified creation}
\usage{
plot_boxplot(
  df,
  x,
  ...,
  ylog = T,
  addPoints = T,
  statTest = T,
  addn = T,
  outputExpr = F,
  colwrap = NA,
  wrapscale = "free",
  methodCompar = "wilcox.test"
)
}
\arguments{
\item{df}{dataframe to analyze}

\item{x}{categorica covariate}

\item{...}{continue value for y abcis. If several values, will create one plot per ... column vs x}

\item{ylog}{If True, set log y scale}

\item{addPoints}{If True, add all points in addition to boxplots}

\item{statTest}{If True, use statistiscal test through ggpubr package}

\item{addn}{If True, add numer of patietn per group}

\item{outputExpr}{if True return the Peccary independant R code instead of the output}

\item{colwrap}{name of a column to wrap the output. Used only if length of ... is equal to 1}

\item{wrapscale}{scale of wrap}

\item{methodCompar}{method to use for statistical test (same as ggpubr)}
}
\description{
Easy creation of a boxplot
}
\author{
Thibaud Derippe (\email{Thibaud.Derippe@gmail.com})
}
