% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Spagh.R
\name{plot_spagh}
\alias{plot_spagh}
\title{Plot Spagh}
\usage{
plot_spagh(
  df,
  x = "TIME",
  y = "DV",
  group = "",
  col = "",
  shape = "",
  lty = "",
  facetwrap = "",
  facetgrid = "",
  group_split = "",
  facetscale = "free",
  median = F,
  point_alpha = 1,
  line_alpha = 1,
  standardisation = "",
  ylog = T,
  xlog = F,
  loq = "",
  filter = "",
  title = "",
  subtitle = "",
  caption = "",
  xlabs = "",
  ylabs = "",
  collabs = "",
  shapelabs = "",
  ltylabs = "",
  colmanual = "",
  shapemanual = "",
  ltymanual = "",
  output = "plot",
  sizetext = 15,
  removeNA = T,
  bkgrd = T,
  filter2 = "",
  bkgrdalpha = 0.3,
  addlines = tibble(type = "none"),
  plotly = F,
  workwithexpr = T
)
}
\arguments{
\item{df}{dataframe to analyze}

\item{x}{x-axis}

\item{y}{y-axis}

\item{group}{grouping variable}

\item{col}{coloring variable}

\item{shape}{shaping variable}

\item{lty}{lty variable}

\item{facetwrap}{variable for facetwrap}

\item{facetgrid}{seecond variable for facetgrid}

\item{group_split}{To create a list of plot}

\item{facetscale}{classical "free", "free_x", "free_y"}

\item{median}{if wanting median to be plotted}

\item{point_alpha}{point alpha (transparancy) value, between 0 and 1}

\item{line_alpha}{line alpha  (transparancy) value, between 0 and 1}

\item{standardisation}{column of number value for standardisation (y = y / standardisation)}

\item{ylog}{True or False}

\item{xlog}{True or False}

\item{loq}{LOQ value}

\item{filter}{main filter}

\item{title}{title}

\item{subtitle}{subtitle}

\item{caption}{caption}

\item{xlabs}{xlabs}

\item{ylabs}{ylabs}

\item{collabs}{col labs}

\item{shapelabs}{ylabs}

\item{ltylabs}{ylabs}

\item{colmanual}{col selection}

\item{shapemanual}{shape selection}

\item{ltymanual}{ylabs}

\item{output}{output}

\item{sizetext}{Sizetext (default 15)}

\item{removeNA}{If True, automatically remove all NA values}

\item{bkgrd}{If True, activate the "background" system}

\item{filter2}{Filter2 (after fixing background)}

\item{bkgrdalpha}{Alpha for background, between 0 and 1}

\item{addlines}{Addline table (for ShinyApp)}

\item{plotly}{True to activate plotly (dynamic plot)}

\item{workwithexpr}{True to use expression (console), false to use character string (ShinyApp)}
}
\description{
Create spaghetti plot with various customisation
}
\author{
Thibaud Derippe
}
