% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readt.R
\name{readt}
\alias{readt}
\title{Reading table with automatic separator guessing and tibble output}
\usage{
readt(
  data,
  na.strings = c(".", "NA"),
  sep = NULL,
  header = T,
  nrows = -1,
  stringsAsFactors = F
)
}
\description{
Reading table with automatic separator guessing and tibble output
}
\author{
Thibaud Derippe (\email{Thibaud.Derippe@gmail.com})
}
