% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecc_search_cov.R
\name{pecc_search_cov}
\alias{pecc_search_cov}
\title{Extract every covariates of a dataset}
\usage{
pecc_search_cov(dataset, ..., returnExp = F)
}
\arguments{
\item{dataset}{dataset to analyze}

\item{...}{grouping columns (ID, occasion,...)}
}
\description{
\code{pecc_search_cov} simplify a dataset to extract only covariates
}
\examples{
pecc_search_cov(Theoph, Subject)
}
\author{
Thibaud Derippe (\email{Thibaud.Derippe@gmail.com})
}
