% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peccary_NCA.R
\name{peccary_NCA}
\alias{peccary_NCA}
\title{Basic Non Compartimental Analysis}
\usage{
peccary_NCA(
  dataset,
  timecol,
  obscol,
  ...,
  BLQcol = 0,
  auc_0_X = 0,
  filter = "",
  keepcov = T,
  plotQC = F,
  outputExpr = F
)
}
\arguments{
\item{dataset}{Dataset to analyze}

\item{timecol}{The name (unquoted) of the time column}

\item{obscol}{The name  (unquoted) of the observation column}

\item{...}{All the names (unquoted) of the grouping column (ID, YTYPE, nAdmin, any covariate....).}

\item{BLQcol}{Does not work properly yet}

\item{auc_0_X}{Numeric input. Compute AUC from time 0 to time auc_0_X}

\item{filter}{Optional filter to be applied on the dataset prior to NCA computation.}

\item{keepcov}{Boolean. If true (default), use \code{\link{pecc_search_cov}} to keep automatically all covariates}

\item{plotQC}{Boolean. If T, create plots to control the output  (default = F)}

\item{outputExpr}{Boolean. If F (default) returns directly NCA results. If T, return R code instead.}
}
\description{
\code{peccary_NCA} compute basic NCA analysis (no extrapolation)
}
\examples{
peccary_NCA(Theoph,Time,conc, Subject, auc_0_X = 12, filter = Subject \%in\% 1:5, plotQC = T)

# To get the Peccary-independant code, add outputExpr = T
peccary_NCA( outputExpr = T, Theoph,Time,conc, Subject, auc_0_X = 12, filter = Subject \%in\% 1:5, plotQC = T)
}
\author{
Thibaud Derippe (\email{Thibaud.Derippe@gmail.com})
}
