% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_correlation.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Correlation simplified creation}
\usage{
plot_correlation(
  df,
  x,
  ...,
  ylog = T,
  xlog = F,
  add = "reg.line",
  cor.coef = T,
  cor.method = "spearman",
  conf.int = F,
  outputExpr = F,
  colwrap = NA,
  wrapscale = "free",
  caption = F
)
}
\description{
Correlation simplified creation
}
\author{
Thibaud Derippe (\email{Thibaud.Derippe@gmail.com})
}
